+++ go.mod
@@ -3,6 +3,8 @@ module github.com/mfcochauxlaberge/karigo
 go 1.12
 
 require (
-	github.com/mfcochauxlaberge/jsonapi v0.6.1-0.20190309204914-844a5c07a848
+	github.com/mfcochauxlaberge/jsonapi v0.6.1-0.20190310040613-85a1d3a48572
 	github.com/mitchellh/copystructure v1.0.0
 )
+
+replace github.com/mfcochauxlaberge/jsonapi => ../../Go/jsonapi
