+++ operation.go
@@ -17,157 +17,50 @@ type Op struct {
 	Value interface{}
 }
 
-// OpAddSet ...
-func OpAddSet(set string) []Op {
-	return []Op{
-		Op{
-			Key: Key{
-				Set:   "0_sets",
-				ID:    "",
-				Field: "id",
-			},
-			Op:    OpSet,
-			Value: set,
+// NewOpSet ...
+func NewOpSet(set, id, field string, v interface{}) Op {
+	return Op{
+		Key: Key{
+			Set:   set,
+			ID:    id,
+			Field: field,
 		},
-		Op{
-			Key: Key{
-				Set:   "0_sets",
-				ID:    set,
-				Field: "name",
-			},
-			Op:    OpSet,
-			Value: set,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_sets",
-				ID:    set,
-				Field: "version",
-			},
-			Op:    OpSet,
-			Value: 0,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_sets",
-				ID:    set,
-				Field: "created",
-			},
-			Op:    OpSet,
-			Value: true,
-		},
+		Op:    OpSet,
+		Value: v,
 	}
 }
 
-// OpAddAttr ...
-func OpAddAttr(set, name, typ string, null bool) []Op {
-	id := set + "_" + name
+// NewOpAddSet ...
+func NewOpAddSet(set string) []Op {
 	return []Op{
-		Op{
-			Key: Key{
-				Set:   "0_attrs",
-				ID:    "",
-				Field: "id",
-			},
-			Op:    OpSet,
-			Value: id,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_attrs",
-				ID:    id,
-				Field: "name",
-			},
-			Op:    OpSet,
-			Value: name,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_attrs",
-				ID:    id,
-				Field: "type",
-			},
-			Op:    OpSet,
-			Value: typ,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_attrs",
-				ID:    id,
-				Field: "null",
-			},
-			Op:    OpSet,
-			Value: null,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_attrs",
-				ID:    id,
-				Field: "set",
-			},
-			Op:    OpSet,
-			Value: set,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_attrs",
-				ID:    id,
-				Field: "created",
-			},
-			Op:    OpSet,
-			Value: true,
-		},
+		NewOpSet("0_sets", "", "id", set),
+		NewOpSet("0_sets", set, "name", set),
+		NewOpSet("0_sets", set, "version", 0),
+		NewOpSet("0_sets", set, "created", true),
 	}
 }
 
-// OpAddRel ...
-func OpAddRel(set, name string, toOne bool) []Op {
+// NewOpAddAttr ...
+func NewOpAddAttr(set, name, typ string, null bool) []Op {
 	id := set + "_" + name
 	return []Op{
-		Op{
-			Key: Key{
-				Set:   "0_rels",
-				ID:    "",
-				Field: "id",
-			},
-			Op:    OpSet,
-			Value: id,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_rels",
-				ID:    id,
-				Field: "name",
-			},
-			Op:    OpSet,
-			Value: name,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_rels",
-				ID:    id,
-				Field: "to-one",
-			},
-			Op:    OpSet,
-			Value: toOne,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_rels",
-				ID:    id,
-				Field: "set",
-			},
-			Op:    OpSet,
-			Value: set,
-		},
-		Op{
-			Key: Key{
-				Set:   "0_rels",
-				ID:    id,
-				Field: "created",
-			},
-			Op:    OpSet,
-			Value: true,
-		},
+		NewOpSet("0_attrs", "", "id", id),
+		NewOpSet("0_attrs", id, "name", name),
+		NewOpSet("0_attrs", id, "type", typ),
+		NewOpSet("0_attrs", id, "null", null),
+		NewOpSet("0_attrs", id, "set", set),
+		NewOpSet("0_attrs", id, "created", true),
+	}
+}
+
+// NewOpAddRel ...
+func NewOpAddRel(set, name string, toOne bool) []Op {
+	id := set + "_" + name
+	return []Op{
+		NewOpSet("0_rels", "", "id", id),
+		NewOpSet("0_rels", id, "name", name),
+		NewOpSet("0_rels", id, "to-one", toOne),
+		NewOpSet("0_rels", id, "set", set),
+		NewOpSet("0_rels", id, "created", true),
 	}
 }
