+++ sources/memory/memory_test.go
@@ -1,4 +1,4 @@
-package sources
+package memory
 
 import (
 	"testing"
