+++ sourcetest/internal/scenarios/example.go
@@ -9,16 +9,24 @@ func init() {
 	Scenarios = append(Scenarios,
 		Scenario{
 			Steps: []interface{}{
-				karigo.OpAddSet("users"),
-				karigo.OpAddAttr("users", "username", "string", false),
-				karigo.OpAddAttr("users", "password", "string", false),
-				karigo.OpAddAttr("users", "created-at", "time", false),
-				karigo.OpAddSet("articles"),
-				karigo.OpAddAttr("articles", "title", "string", false),
-				karigo.OpAddAttr("articles", "content", "string", false),
-				karigo.OpAddAttr("articles", "created-at", "string", false),
-				karigo.OpAddAttr("articles", "updated-at", "string", false),
-				karigo.OpAddRel("users", "articles", false),
+				// Users
+				karigo.NewOpAddSet("users"),
+				karigo.NewOpAddAttr("users", "username", "string", false),
+				karigo.NewOpAddAttr("users", "password", "string", false),
+				karigo.NewOpAddAttr("users", "created-at", "time", false),
+
+				// Articles
+				karigo.NewOpAddSet("articles"),
+				karigo.NewOpAddAttr("articles", "title", "string", false),
+				karigo.NewOpAddAttr("articles", "content", "string", false),
+				karigo.NewOpAddAttr("articles", "created-at", "string", false),
+				karigo.NewOpAddAttr("articles", "updated-at", "string", false),
+
+				// Relationships
+				karigo.NewOpAddRel("users", "articles", false),
+
+				// Data
+				karigo.NewOpSet("users", "", "id", "abc123"),
 			},
 			Verif: map[string][]jsonapi.Resource{},
 		},
