+++ transaction.go
@@ -8,47 +8,47 @@ func TxNotImplemented(snap *Snapshot) {
 	snap.Fail(ErrNotImplemented)
 }
 
-// TxUpdatePlayer ...
-func TxUpdatePlayer(snap *Snapshot) {
-	id := snap.Res.GetID()
-	score := snap.Res.Get("score").(int)
+// // TxUpdatePlayer ...
+// func TxUpdatePlayer(snap *Snapshot) {
+// 	id := snap.Res.GetID()
+// 	score := snap.Res.Get("score").(int)
 
-	// snap.LockAll()
-	// snap.Lock("players")
-	// snap.Ready()
+// 	// snap.LockAll()
+// 	// snap.Lock("players")
+// 	// snap.Ready()
 
-	snap.Add(Op{
-		Key: Key{
-			Set:   "players",
-			ID:    id,
-			Field: "score",
-		},
-		Op:    OpSet,
-		Value: score,
-	})
+// 	snap.Add(Op{
+// 		Key: Key{
+// 			Set:   "players",
+// 			ID:    id,
+// 			Field: "score",
+// 		},
+// 		Op:    OpSet,
+// 		Value: score,
+// 	})
 
-	snap.Flush()
+// 	snap.Flush()
 
-	ranks := snap.Collection(QueryCol{
-		Set:    "players",
-		Fields: []string{"id"},
-		Sort:   []string{"-score"},
-	})
+// 	ranks := snap.Collection(QueryCol{
+// 		Set:    "players",
+// 		Fields: []string{"id"},
+// 		Sort:   []string{"-score"},
+// 	})
 
-	for i := range ranks {
-		snap.Add(Op{
-			Key: Key{
-				Set:   "players",
-				ID:    ranks[i]["id"].(string),
-				Field: "rank",
-			},
-			Op:    OpSet,
-			Value: i + 1,
-		})
-	}
+// 	for i := range ranks {
+// 		snap.Add(Op{
+// 			Key: Key{
+// 				Set:   "players",
+// 				ID:    ranks[i]["id"].(string),
+// 				Field: "rank",
+// 			},
+// 			Op:    OpSet,
+// 			Value: i + 1,
+// 		})
+// 	}
 
-	// snap.Release("players")
-	// snap.ReleaseAll()
+// 	// snap.Release("players")
+// 	// snap.ReleaseAll()
 
-	snap.Commit()
-}
+// 	snap.Commit()
+// }
